feat: add `--before` to allow filtering to posts before a given date
This can be useful to retroactively generate a webring as if it was done
at a previous time.

The caveat it that it is common for RSS feeds to only contain the most
recent N posts. So, if `--before` date is too far in the past, it is
likely that the feeds will not have any articles from before that time,
and openring will come up empty.