refactor!: switch to `feed-rs`
This library has a simpler API that unifies Atom and RSS, and is more
maintained. As a side effect, it also allows us to resolve our `chrono`
security advisories.

Unfortunately, since we are no longer responsible for parsing the date,
we cannot support invalid, naive dates. However, there is a ticket
for better support upstream, and upstream already handles a much wider
variety of issues that we did.

BREAKING CHANGE: We no longer can parse feeds whose dates do not include
  timezones. Support for this will need to be fixed upstream.

Ref: https://github.com/feed-rs/feed-rs
Ref: https://github.com/feed-rs/feed-rs/blob/master/feed-rs/src/parser/util.rs
Ref: https://github.com/feed-rs/feed-rs/issues/182
Signed-off-by: Luke Hsiao <luke@hsiao.dev>