feat: add caching options to respect headers
We want to respect `Etag` and `Last-Modified` headers when sending
requests to reduce resource strain on the servers providing feeds.
Simlarly, we want to respect `Retry-After` if a server provides
that header when responding with an HTTP 429.

This patch respects both by introducing a local cache option in
`.openringcache`, which is a simple CSV file with the schema: url,
timestamp, retry_after, last_modified, etag, and body, where body is the
entire content of the response body last time we fetched the feed.

With this local cache, if we have a value for Retry-After, we know we
were throttled, so we skip sending a request and just use the feed from
the cache.

Otherwise, if we have a cache value, we send a conditional request,
setting `If-Modified-Since` and `Etag` headers in the request.

If we don't have a cache value, we send an unconditional request.

The implementation is modeled after the caching in lychee [[1]].

Tested: Mostly tested manually on rachelbythebay.com, since this
  is the only feed I'm personally aware of that has an opinionated and
  detailed documentation for how it behaves.

  From manual testing, I see we properly skip sending requests when
  we were 429'ed previously, and that we just use the cache value
  when we get a 304.

[1]: https://github.com/lycheeverse/lychee
Ref: https://rachelbythebay.com/w/2023/01/18/http/