perf: switch from `rayon` to `tokio`
`openring` is I/O bound, not CPU bound. With `rayon`, we would only
fetch N feeds at a time, where N is the number of cpus available on
the machine. Then, we would wait for a feed to respond before sending
the next one.

This is highly inefficient.

With `tokio`, we instead concurrently fetch _all_ the URLs. This is
a significant speedup for those with long lists of feeds or low numbers
of cores.